THE Window class + OpenAL header reorganize
