+++ src/include/sndx/audio.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "./audio/al.hpp"
+
+#include "./audio/audio_decoder.hpp"
+#include "./audio/mp3.hpp"
+#include "./audio/vorbis.hpp"
+#include "./audio/wav.hpp"
